smc_trading_agent/data_pipeline/ingestion.py
import faust
from kafka import KafkaProducer
import pandas as pd
import numpy as np

class MarketDataProcessor:
    def __init__(self):
        # self.kafka_producer = KafkaProducer()
        # self.faust_app = faust.App('smc-processor')
        pass
        
    async def process_tick_data(self, tick_data):
        # Real-time OHLCV construction
        # Volume profile analysis  
        # Liquidity zone detection
        pass

    def get_latest_ohlcv_data(self, symbol: str, timeframe: str, limit: int = 100) -> pd.DataFrame:
        """
        Simulates fetching the latest OHLCV data for a given symbol.
        In a real application, this would fetch from a database or a live feed.
        """
        end_time = pd.Timestamp.now(tz='UTC')
        start_time = end_time - pd.to_timedelta(limit, unit='h') # Assuming 1h timeframe for simulation
        
        timestamps = pd.to_datetime(np.linspace(start_time.value, end_time.value, limit))
        
        # Simulate some price action
        price_path = 60000 + np.random.randn(limit).cumsum() * 10
        
        data = {
            'timestamp': timestamps,
            'open': price_path,
            'high': price_path + np.random.uniform(0, 50, size=limit),
            'low': price_path - np.random.uniform(0, 50, size=limit),
            'close': price_path + np.random.randn(limit) * 5,
            'volume': np.random.uniform(10, 200, size=limit)
        }
        df = pd.DataFrame(data)
        # Ensure high is the highest and low is the lowest
        df['high'] = df[['high', 'open', 'close']].max(axis=1)
        df['low'] = df[['low', 'open', 'close']].min(axis=1)
        
        return df
