smc_trading_agent/requirements.txt
# Core ML/AI
tensorflow>=2.10.0
torch>=1.12.0
stable-baselines3>=1.6.0
numba>=0.56.0
scipy>=1.9.0
scikit-learn>=1.1.0
gymnasium>=0.26.0

# Data Processing
pandas>=1.5.0
numpy>=1.21.0
PyYAML>=6.0
pandas-ta>=0.3.14b
vectorbt>=0.25.0
pydantic>=2.0.0
python-dotenv>=1.1.0

# Financial Analysis
quantstats>=0.0.61

# Streaming & Messaging
faust-streaming>=0.10.15 # Using the maintained fork
kafka-python>=2.0.2
redis>=4.3.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0

# Trading APIs
ccxt>=2.0.0
websockets>=10.4
aiohttp>=3.8.0

# Monitoring
prometheus-client>=0.15.0
grafana-api>=1.0.3
python-json-logger>=2.0.7

# Testing
pytest>=7.1.0
pytest-asyncio>=0.20.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0

# Development Tools (for exported-assets scripts)
plotly>=5.0.0
xgboost>=1.6.0
